Decouple StateMachine from ExtendedEventHandler.
Instead, inject it where it is needed.

This is possible because StateMachines always emit events to a
StoredEventHandler associated with SkyKeyComputeState (driven by the
Skyframe requirement that SkyFunctions replay events if there is a restart).

This simplifies some error handling code by making the ExtendedEventHandler
immediately available to error handling callbacks, which would otherwise
need to propagate errors upwards.

PiperOrigin-RevId: 548175834
Change-Id: Ie384bb05561d6503513603971191d13b545bce18