Have `AspectValue` extend `ConfiguredAspect` and use a separate subclass if transitive packages are tracked.
`ConfiguredAspect` has two fields: actions and providers. The actions already need to be copied over to `AspectValue` in order to support clearing the providers, meaning that `ConfiguredAspect` is really just stored for the providers.

Using inheritance isn't ideal but makes it easier to have a single `ProviderCollection` implementation and avoids needing to refactor the many places that expect a `ConfiguredAspect` and are now getting an `AspectValue`. There is still value in constructing the intermediate `ConfiguredAspect` instances instead of taking actions and providers individually - we verify that they went through `ConfiguredAspect`'s builder, which enforces some business logic.

PiperOrigin-RevId: 548265903
Change-Id: I99252acd19a2e00ff366073f770433cf02bf23b1