feat: Add `additional_linker_inputs` option to `cc_library` rule
This is achieved by the following related changes:

1) Moving the `additional_linker_inputs` attribute up from the `cc_binary_base`
rule to the `cc_rule`. The rationale here is that any rule that allows `linkopts`
to be set should also support `additional_linker_inputs`.

2) Adding the `additional_linker_inputs` attr to the `cc_library` rule context.
This was copied verbatim from the existing `cc_binary` attribute list.

3) Appending the `ctx.files.additional_linker_inputs` list to the list of
linker scripts wherever they are passed as `additional_inputs` in the `cc_library.bzl`
builtin file corresponding to the `cc_library` rule.