Allow CppLinkAction to inspect input sizes in resource estimation
This PR instrumets the CppLinkAction to print log lines for every link
operation and dumps the number of inputs, the aggregate size of the
inputs, the predicted memory usage, and the actual consumed memory.

Calculating the aggregate size of the inputs is a costly operation
because we need to expand a nested set and stat all of its files, which
explains why this PR is complex.  I've modified the local memory
estimator to compute the inputs size exactly once per link as we need
this information both to predict memory consumption _and_ to emit the
log line after the spawn completes.

This is meant to aid in assessing the local resources prediction for
linking with the goal of adjusting the model, but that should be done
in a separate change.

Helps address #17368.