Print timestamp at the start of test execution.
The intent is to simplify debugging of timeouts and pre-test latency: this way for timeouts there will be a clear measured interval between `Executing tests...` to `Test timed out at...` lines.

One side-effect of this will be to guarantee all test logs are unique, rather than just the ones that contain a timestamped log entry or timeout. I don't expect this to be an issue, as the vast majority of logs should already be unique for some reason or another.

RELNOTES: Print timestamp at the start of test execution logs.
PiperOrigin-RevId: 549006113
Change-Id: Iae7441a60bd7f6cbca8e96369618a44f68273039