Don't store `CHECK` reverse dep operations on an already-done node.
These operations are accumulated in a list of "data to consolidate". Consolidation happens when either 1) the node finishes or 2) a threshold size of data to consolidate is met (currently set to # of existing rdeps).

When a node is already done, it is likely that consolidation never happens (the threshold is never met). Furthermore, consolidation of `CHECK` ops on an already-done node is a no-op besides preconditions checks.

PiperOrigin-RevId: 549142409
Change-Id: Ibc7a50e7ee7b1e9bc04a24be8b36d12f7fe3984b