[Skymeld] Update `test_test#test_build_fail_terse_summary`.
For Skymeld, it possible for the execution error to be hit before any of the
test targets' analysis is done, therefore failing the test. Example failure:
[log](https://storage.googleapis.com/bazel-untrusted-buildkite-artifacts/01896905-0d3e-4319-b834-6541a70b722a/src/test/shell/integration/test_test/test_attempts/attempt_1.log).

This CL adds `--build_tests_only` to the test to make sure that by the time we
hit the execution error, some test analysis is already done.

PiperOrigin-RevId: 549278633
Change-Id: Ia31e402d16ae64716e80e5f1f735a88b61bb2b5e