Prevent non-deterministic test log output in test-setup.sh.
When test-setup.sh kills the process group for the "cleanup helper" subshell (`kill_group SIGKILL $cleanupPid`), the `sleep` command running inside the subshell gets killed first. It appears that, if enough time elapses until the subshell is itself killed, it will get a chance to print a "XXXXX: Killed: 9    sleep 10" message, which in addition to being printed non-deterministically, embeds a non-deterministic PID. Therefore, this subshell's output (which is never useful) should be redirected away from the test log.

I suspect this to be the root cause for the flakiness of test_exclusive_tag in remote_execution_test.sh, at least on MacOS where it's extremely likely to occur. I haven't succeeded in reproducing it on Linux, so there could be a separate issue.

PiperOrigin-RevId: 549303669
Change-Id: I0ebcc708e05250222d3faa32a6f889ccfebefc78