Make native `JavaOutput` a `StructApi` to allow both native and Starlark instances to be added to a `depset`
`depset` type-checking relies on `Starlark.type()` which finds the lowest ancestor class with a `@StarlarkBuiltins` annotation. For Starlark providers this is `StructApi`, so we have to make it the same for our native class.

This is needed so we can de-duplicate `JavaOutput` instances efficiently in Starlark (otherwise we run into https://github.com/bazelbuild/bazel/issues/17170).

Also updates `JavaOutput` translation in `JavaPluginInfo` that was missed in an earlier change.

PiperOrigin-RevId: 549520899
Change-Id: I7811c57bbbce41ab92b50938889d4e2c1b8cf34b