Refactor digest for Starlark-defined rules
We now store the transitive digest of the .bzl currently being initialized in the BzlInitThreadContext object. This means that, when determining the digest for a newly created Starlark rule class, we no longer need to use stack inspection (Debug#getCallStack) to find the outermost module.

(Originally I misread the code and thought I was fixing a semantic bug wherein we would use the innermost module's digest instead of the outermost's. But that was fixed in unknown commit)

The digest remains incorrect for rule classes generated by analysis_test. Added TODO/comments.

PiperOrigin-RevId: 549626857
Change-Id: I8c11c05515e3cb47c24e3277d3a74ef496c27d42