Add private API to convert Starlark `JavaInfo` instances back into native
This is needed to avoid a memory regression with using Starlark `JavaInfo` instances in `java_common.merge` and `java_common.compile`.

This is a stop-gap to start using the Starlark implementations and delete a fair bit of native code. When we figure out how to deal with the memory regression, this can be deleted.

PiperOrigin-RevId: 550489669
Change-Id: I6fdbd240291c804b7de94f6124e03c90bbe597b4