Implement collocated allowlist checking on java, cc and py_proto_library
Previous implementation didn't work, because the proto_common.compile is called from an aspect, where owner is the same as underlying proto_library.

The label of lang_proto_library is not available to aspect. Because of this a second call is needed and it needs to be called from each lang_proto_library.

PiperOrigin-RevId: 550584299
Change-Id: I77b8de5ac8ad636cfd41181dd017dd27de1fbbb1