Generate combined coverage artifacts after exclusive tests have run.
The combined coverage report requires tests to be executed in order to
be built. If it is requested as part of the build then those tests will
be executed.

This fixes SkyframeBuilder and SkyframeBuildView to trigger generation
of the combined report (or all coverage artifacts in the case of the
latter) only after all exclusive tests have been run.

Fixes https://github.com/bazelbuild/bazel/issues/6005

PiperOrigin-RevId: 550869458
Change-Id: I67d28184e59618a756b0114bf77996e3e9c26ffb