Support aliases inside Automatic Exec Groups
It is possible to use an alias to declare a toolchain, which would be an issue with AEG since creation of AEGs happens before toolchain types are resolved.

By design, name of an AEG (Automatic Exec Group) is mapped to a toolchain type registered on a rule. If a rule registered toolchain which is an alias, that's causing issued since that toolchain will be resolved after the creation of AEG. In that case we could have something like:
`exec_group(name)

I also modified `AndroidSdkProvider` to take correct toolchain when AEGs are enabled (not from the `default-exec-group`).

PiperOrigin-RevId: 550902606
Change-Id: Ied4525095616ddeec992d16e20911188c3d866a1