Fix error message for native Java providers
For some native providers (like `JavaInfo`), we were forced to remove the `@StarlarkBuiltin` annotation that allowed `Starlark#type` to find the correct name. This was so that we could add both Starlark and native types to the same `Depset`.

This means, we need to get the type differently in the error message if provider validation fails, otherwise we just get `struct`

PiperOrigin-RevId: 551138227
Change-Id: Iabf47b6eac7dc33f82a2cf8decc7689fe83dcfc1