Enforce a precondition that the argument to FileSystem#resolveSymbolicLinks is an absolute path.
Filesystems have no concept of a working directory, so resolving a relative path cannot produce a meaningful result. Note that only other filesystems may call FileSystem#resolveSymbolicLinks; filesystem users must call the method with the same name on a Path, which is already guaranteed to be absolute (as enforced by the Path constructor).

PiperOrigin-RevId: 551530186
Change-Id: Ib972f3f3928043ad1213343405eeff4c117e840f