Exclude ArtifactNestedSetKeys from the path to cycles.
ArtifactNestedSetKeys are just an implementation detail of Skyframe. The user
only needs to know about the relationship between an artifact and an action, and
not the particular ArtifactNestedSets in between.

This also fixes a crash where we try to prettyPrint ArtifactNestedSetKeys. This
was just an inconsistency between skipping a key type and allowing which key
types to be printed: ANSK was not skipped, but not in the allowlist to be
printed either.

PiperOrigin-RevId: 551840565
Change-Id: Ie1257155a0e2821d2d32e8147b71dcfba3f7dd82