Fix crash in Aspect error handling.
The crash occurs when an aspect's dependency is in error. depValueMap
will always be null in that case so any errors propagating via
transitiveRootCauses aren't correctly handled.

This crash is relatively rare because computeDependencies usually doesn't
error as much of the time, it simply repeats the calculation performed
by the base configured target. However, if the Aspect has its own attribute
labels and those are in error, the result is a crash.

PiperOrigin-RevId: 551923788
Change-Id: If3820fe4f59d82e99bf1c477f95dc5675ae7680a