Rename PROCESS_TIME to REMOTE_PROCESS_TIME
to reflect the fact that it's used to mark remote execution wall time.

Currently, PROCESS_TIME is used in LocalSpawnRunner to record the local execution time, but displayed as remote execution time in the Profiler which is confusing. This change reduces the chance PROCESS_TIME is misused in the future and a following change will fix LocalSpawnRunner.

PiperOrigin-RevId: 552467817
Change-Id: I4383be640571fd816fb2c71fec29f60e693d4347