[Skymeld] Fix the flaky `testTwoConflictingTargets_keepGoing_behaviorDifferences`
Consider 2 targets `//foo` and `//foo/bar` which generates 2 artifacts `foo` and
`foo/bar`.

In skymeld and `-k` mode, it's possible that either 1 of the 2 artifacts are
built or none at all. The last case can happen when the
`ArtifactPrefixConflictException` are thrown for both the evaluation of
`BuildDriverKey(//foo)` and `BuildDriverKey(//foo/bar)`, which is possible given
that the `IncrementalArtfiactConflictFinder` can be accessed from multiple
threads.

This CL updates the test to reflect that.

PiperOrigin-RevId: 552476900
Change-Id: Ifc67a9f1c09ab1887656ea491a6f2247484163e2