Resolve symlinks correctly even when they cross filesystems.
Instead of delegating RemoteActionFileSystem#resolveSymbolicLinks to the local filesystem, use its base implementation together with a custom implementation of FileSystem#resolveOneLink which is able to stitch together the local and remote filesystems.

In addition, have every filesystem method that needs to resolve symlinks symlinks call resolveSymbolicLinks explicitly (as delegating to the underlying filesystems would cause cross-filesystem symlinks not to be followed correctly).

The end goal of these changes is to let SkyframeActionExecutor traverse the RemoteActionFileSystem normally when collecting metadata for non-symlink outputs materialized as symlinks, instead of relying on the latter to provide pre-constructed metadata. This is only a step on the way there; the way SkyframeActionExecutor collects metadata remains unchanged for now.

PiperOrigin-RevId: 552743574
Change-Id: I5921ded8f7b9c51c31406daf0441ee3c2e17308b