Add the cherry-picker-on-close.yml and cherry-picker-on-comment.yml
This is for the cherry-pick automation.

cherry-picker-on-close.yml will be triggered when a PR is merged and closed by copybara. cherry-picker-on-comment.yml will be triggered when a new issue is milestoned usually by bazel-io with a comment, "bazel-io fork X.Y.Z".

Once triggered, the continuous-integration/actions/cherry-picker/index.py will run to:
1) Check if the commit was made by copybara
2) Check if PR was closed
3) Check if PR was approved
4) And cherry-pick the commit into proper release branch. If not cherry-pickable, then the bot will notify in the milestoned issue
5) Create a PR

PiperOrigin-RevId: 553105331
Change-Id: Ib52a56e49e58ea2adf92c8398ad07dda06eb9073