Announce an `ExecRequest` event for run commands on the BEP.
Also introduce `--experimental_bep_exec_request_include_residue` to optionally include the residue of the run command in said `ExecRequest`.

Currently the client on `bazel run` commands uses the `ExecRequest` returned via the grpc stream from the server to run the built executable. With this change, the BEP is an alternative location to observe the `ExecRequest`.

`bazel run` commands now delegate to `RunBuildCompleteEvent` for `BuildCompleteEvent` information - this allows the `ExecRequestEvent` to be sent before `BuildCompleteEvent`.

PiperOrigin-RevId: 553159588
Change-Id: I166a75ddefdbe6101a4d0bd76b6cd4b29aa2bcc0