Use a `ConstantVersion` for non-incremental evaluations.
This is a prerequisite for not storing a version field in edgeless nodes. In order to satisfy a check in `ParallelEvaluator#informProgressReceiverThatValueIsDone`, the node and the evaluator must agree on the version.

Also skip calling `deleteDirty` if incremental state is not being tracked, since it doesn't make sense with a non-`IntVersion` and is irrelevant without incrementality anyway.

PiperOrigin-RevId: 553489413
Change-Id: Iedd5ad22e99ad3ef7c9398b2105ff289e1fd00b8