Remove deploy jar Java runtime dependency for non-executable targets
This allows deploy jars for `java_binary` targets with `create_executable = False` to compile for target platforms for which no standalone Java runtime is available (e.g. Android).

Along the way, this removes a redundant check (`runtime.hermetic_files` is never `None`) and ensures that files coming from the hermetic runtime are only added if `hermetic` is set to `True`.

Work towards #17085

Closes #19140.

PiperOrigin-RevId: 553855531
Change-Id: I56ce730190498e2adb2f8acba709f02ec9c13ef4