Exempt exec transitions from "banned flags" list.
Starlark transitions generally can't change a certain class of flags:
https://github.com/bazelbuild/bazel/blob/d0620d1331f4dc86dc057c76be71001b6b877276/src/main/java/com/google/devtools/build/lib/rules/config/ConfigGlobalLibrary.java#L118-L137

Exec transitions need to be exempt from this. This is because exec
transition logic predates Starlark transitions and already sets a lot of these
flags. Maybe one day we can refactor away those dependencies. But that's what the
current semantics require.

The implementation is unfortunately hacky: any transition defined in a `*_exec_platforms.bzl` file is considered an exec transition. See the TODO for ideas on how to consider this more principally.

PiperOrigin-RevId: 554959721
Change-Id: I83986c6c0a9855fe9f15d7ce78fc1e3cf4c5861a