Remove unused Apple simulator starlark API
As far as I can tell these starlark APIs have never been used by rules_apple, and I can't find any other uses of them on GitHub. The intent of these was to replace the iOS specific API which is used in rules_apple but at this point if these are ever needed by the rules we should use a custom starlark flag instead of a bazel flag.

Closes #19169.

PiperOrigin-RevId: 555304875
Change-Id: I04f81cb940b97fb855ee4c26b40bb17217d9a417