Move the shouldTrustRemoteArtifact check to validateOutputs.
This is necessary, when building without the bytes, to correctly handle the case where a toplevel action A materializes an output as a symlink to an artifact produced by a non-toplevel action B and B's output (but not A's) is deleted from the output tree before incrementally rebuilding. In this case, when the ActionCacheChecker is asked whether A should rerun, the ActionMetadataHandler will follow the symlink from the local filesystem into the remote output tree (*) and yield remote metadata for action A, which must not be trusted so that A reruns and causes B's output to be downloaded again.

(*) This currently only occurs in the symlink-to-file case, because a symlink-to-directory is not correctly followed by the RemoteActionFileSystem (although it will soon be). Both cases ought to have been covered by the BwoB integration tests, but a typo (now fixed) caused coverage for the symlink-to-file case to be missing.

PiperOrigin-RevId: 555918488
Change-Id: I97b37c4257cf26e6cb9957f8c5663a9637209049