Implement `NonIncrementalInMemoryNodeEntry` separately from `IncrementalInMemoryNodeEntry`.
This reduces the shallow heap of `NonIncrementalInMemoryNodeEntry` from 40 to 24 bytes, which is especially good because users often opt for non-incremental builds to save memory. The overall savings are somewhat lessened by existing optimizations which already reduce the number of retained nodes in non-incremental builds:

* `--heuristically_drop_nodes` which discards nodes on the fly.
* `GlobbingStrategy#NON_SKYFRAME`, which is used to reduce the number of nodes in the glob subgraph on non-incremental builds since we don't need incremental correctness.

Some of the state checks for the "force rebuild" lifecycle are loosened to allow `NonIncrementalInMemoryNodeEntry` to act as if it has never been built before, as opposed to supporting some special incremental building state only for rewinding. I left a couple `TODO`s to tighten this up as I work on incremental rewinding support.

PiperOrigin-RevId: 555934542
Change-Id: Ie9600706c7f34b41a11401812534571e516e5a3b