Rename erroneous references to the "mainRepositoryName" to the actual term "workspaceName"
A `RepositoryName` object is always a canonical name. So a `RepositoryName mainRepositoryName` should just always be `RepositoryName.MAIN`, which corresponds to the empty string. What is actually meant in these files is the _workspace name_ (i.e. `workspace(name="X")`, `WorkspaceNameValue`, etc), alternatively known as the "execpath/runfiles path prefix", which incidentally is somewhat of an obsolete concept in the Bzlmod world (just the hardcoded string `"_main"`).

PiperOrigin-RevId: 557383945
Change-Id: If479c71b34aa396987c77eaa148d29a3636b0b9d