Merge java_binary/java_test with launcher into a single rule
Add `//tools/build_defs/java:launcher_label_alias` target that reads native java_launcher/host_java_launcher flag. The target will be later replaced with label_flag. The implementation returns no providers, when flag is unset. Later on this can be solved, by implementing a "sentinel" launcher, that's used when the flag is unset. `label_flag` acts as a proper `alias`.

Use `//tools/build_defs/java:launcher` as default value for `launcher` field in Java rules, remove computed default. This replicates the previous behavior. First the value of `launcher` attribute is used if set. Then the value of the flag is used. If the flag is also not set, the default is used.

Handle `launcher = None` in the macro part. attr.label has a weird behaviour: If set to None, the default value is used (and that is a flag).

Drop additional flavours of java_binary/java_tests. They are not needed.

Update mocks for unit and integration tests

There are further cleanup of the code possible, however probably better to do them in a separate change.

PiperOrigin-RevId: 557481404
Change-Id: I581e790aad8187b02d161ee6c0d76108fa5e8261