Do not rerun module extensions when only imports or locations change
When using the lockfile, extension evaluation results are now persisted even if the imports (`use_repo`s) fail validation against the actually generated repos and reused if only imports or Starlark locations change.

This requires storing `ModuleExtensionMetadata` in the lockfile to ensure that the correct fixup warnings are shown if imports are updated but the extension isn't rerun.

Closes #19253.

PiperOrigin-RevId: 557878432
Change-Id: I25909294b118fb445f9c48f61e18463762f78208