Obey --experimental_inprocess_symlink_creation in RunfilesTreeUpdater.
The bulk of this CL is just plumbing the flag from the configuration into the RunfilesSupplier implementations. The functional changes are in RunfilesTreeUpdater; compare with the preexisting SymlinkTreeStrategy implementation.

Additionally, I've fixed some variable and method names to exactly match the flags defined in CoreOptions; the code is less confusing and easier to refactor that way.

The motivation behind this change is to investigate the feasibility of defaulting to the in-process implementation and deleting the out-of-process one, which has known issues (see https://github.com/bazelbuild/bazel/issues/4327).

PiperOrigin-RevId: 558112830
Change-Id: I4c509be665776013ffc2d483694df515d5ab32f2