Create separate class for Aspect keys with base aspects
Extend `AspectKey` with 2 classes: `SimpleAspectKey` for aspects not depending on other aspects and `AspectKeyWithBaseAspect` for aspects depending on one or more base aspects. This is for memory optimization as in most cases the aspect will not depend on other aspects and the `baseKeys` list will be empty and there is no need to hold a reference to it.

Also replace `Objects.hashCode` with `HashCodes.hashObjects`

PiperOrigin-RevId: 558245545
Change-Id: Iea8dc3ab71cc7f72dc591136a76f92b3b3831527