Allow customizing errors during remote connections
Almost all remote caching and execution services are protected by some
kind of authentication or networking requirements (like VPN).  These
vary from deployment to deployment so Bazel cannot print useful error
messages on its own to guide users to a solution when they experience
a connectivity problem.

To resolve this, add a new --experimental_remote_init_failure_message
flag that provides a template for the error messages that Bazel prints
when the connection to the build farm fails.  The default value respects
the existing messages to not change behavior, but users can tweak its
value to provide guidance to users.

The idea is that the .bazelrc of a repo will customize this flag to
provide a useful message in the context of that repo.