Store all `javacopts` as a `NestedSet` instead of `List`
Unfortunately, this is not as straightforward as just changing the container type. javac options are a little special, in that we allow users to specify them in as arbitrary strings and we are expected to tokenize them.

However if we naively store the list of options in the `NestedSet`, upon flattening the set to get back the list, we lose duplicate occurrences. This gets particularly tricky with `-source` and `-target` which don't have an equivalent `--arg=val` form that we could migrate to. These arguments _must_ be passed to `javac` as separate tokens.

So to get around this, we de-tokenize the list of options into a single space-separated string, and re-tokenize when required as a list.

For backwards-compatiblity, the existing list returning methods are kept as is and  I've added a new method that returns the `NestedSet`/`Depset`. In some cases, the return type is chosen by a boolean `as_depset` parameter.

The plan is to release this change, migrate the depot to using `depset`s, and then drop the list methods / change default behavior.

`JavaPackageConfigurationProvider` might be the one place we may have to keep storing both the list and set instances, because that is a _very_ hot code path and tokenizing on-demand would be too expensive.

PiperOrigin-RevId: 559676174
Change-Id: Iee880bfee5328033e20018a1f99d63e6fac2ee54