Use "minSdkVersion" input from values for empty manifest
By default `AndroidManifest.asEmpty()` uses inputs of `""` for package and `"1"` for the the min sdk, this is causing the merger validation to later complain that the min sdk is being overridden by the manifest values sdk version (in `XmlDocument.verifyVersion`). Of course the input sdk version is synthetic and was not declared by the user so this error is spurious.

PiperOrigin-RevId: 559757030
Change-Id: I14d18905f32bafc0ccd6aa4f3dec48415db7df74