Add back `synchronized` to `NonIncrementalInMemoryNodeEntry#resetForRestartFromScratch`.
This was accidentally removed in https://github.com/bazelbuild/bazel/commit/20c541cf2e910d9e85ccd42d076063fba97251a7 when it was copied over from `IncrementalInMemoryNodeEntry`. As far as I can tell, it's actually safe without it because of the way that Skyframe uses it (and it's been released without issue). However, it's much more obviously correct with `synchronized`, and restarts are expected to be rare so there's no performance concern.

PiperOrigin-RevId: 559768635
Change-Id: Idb25f9daeed83073ff8fda2dc47355b03f7d501f