Clean up code related to `NodeEntry#resetForRestartFromScratch`.
Improve documentation to note the relationship to rewinding and correct an important specification about erasing temporary direct deps but persisting rdeps. Add a unit test to `NonIncrementalInMemoryNodeEntry`.

Remove code supporting `resetForRestartFromScratch` with incrementality. It is currently unused, and will need a different approach to work with rdeps. Also remove the bit of test coverage in `MemoizingEvaluatorTest` in favor of starting from scratch.

PiperOrigin-RevId: 559781645
Change-Id: Ib3a850c173124b6a157cc1353dc3d1918faa031b