Add limited support for `resetForRestartFromScratch` on `IncrementalInMemoryNodeEntry`.
Introduces a new method `NodeEntry#getResetTemporaryDirectDeps`. When graph edges are kept, nodes are responsible for tracking direct deps that were registered but cleared by a reset. This will be used to prevent registration of duplicate reverse deps (not part of this change). Alternatively, we could actually remove reverse deps on a reset, but this approach is cleaner as all of the behavior is implemented on the resetting node (not dependency nodes) and the rdep edges would need to be added back anyway.

This is the first bit of new behavior to support rewinding with incrementality. The implementation is currently limited to incremental nodes on their initial build - I intend to make each change small to keep it easier to understand, but will leave `TODO`s where I know that more support is needed.

PiperOrigin-RevId: 560131483
Change-Id: I8eb038fec880f56620810de0b310772b55e81fce