Ensure lockfile is updated after reset to pre-build state
The events that update the lockfile are important for incremental
correctness and thus need to return `true` from `storeForReplay`.

Before this change, if a build creates the lockfile because it didn't
exist and the lockfile is then deleted, subsequent builds did not
regenerate it as the relevant SkyFunctions wouldn't rerun and the events
were not replayed.