Stop injecting metadata for non-symlink outputs materialized as a symlink (to a preexisting remote artifact) in RemoteActionFileSystem.
Instead, let ActionMetadataHandler traverse the filesystem to collect the metadata, including the materialization path. This is a better separation of concerns and extends this possibility to cases where the symlink is created out of process (see newly added integration tests), which is required to maintain parity between with and without the bytes builds.

PiperOrigin-RevId: 560689494
Change-Id: I3e5f0ae87db32832dc5a830c016a9cd1f18f9659