Use starlark_doc_extract as a source for rules docgen in the build encyclopedia
As we are moving rules from native to Starlark-based @_builtins, we should move
rule docs from comments in Java source files to the new .bzl files where the
rules are implemented.

Note that the current build encyclopedia pipeline does not perform any Markdown
rendering, so (for now) migrated rule docs will still need to me marked up in
plain HTML, not in Markdown.

The change necessitated some refactors to the BE docgen tool:

* RuleDocumentationAttribute now stores its fields inside itself instead of
  retrieving them from an Attribute object - since those fields may now come
  from an AttributeInfo proto instead. As a side effect, this allows us to make
  RuleDocumentationAttribute immutable.
* We needed a way to render links to .bzl rule sources, which we get in the
  form of labels instead of file paths. So instead of the previous hacky method
  of searching for "io_bazel" in input file paths (which was utterly broken on
  RBE and Windows in addition to general fragility), we now have a
  SourceUrlRetriever for transforming file paths and labels (including labels
  in @_builtins repo) into source code URLs.

PiperOrigin-RevId: 561088516
Change-Id: I5ec76b87db56d2cc51a9c8bd9e063428a42cbd3a