Ensure lockfile is updated after reset to pre-build state
The events that update the lockfile are important for incremental correctness and thus need to return `true` from `storeForReplay`.

Before this change, if a build creates the lockfile because it didn't exist and the lockfile is then deleted, subsequent builds did not regenerate it as the relevant SkyFunctions wouldn't rerun and the events were not replayed.

Closes #19343.

PiperOrigin-RevId: 561287438
Change-Id: I549f99b896a0095e8ffc35b7bacc8a841a44219a