Fix precompiled lib symlinks missing in runfiles of cc_shared_library
26d882f92943587d39460320694b17865ea74f91 was intended to remove unnecessary files from runfiles
However it also removes (some?) required symlinks from runfiles.
E.g. libtensorflow_cc.so.2.13.0 is present in the runfiles but
`libtensorflow_cc.so.2` is not leading to failures as programs are
linked to the latter.

Use the same logic used in other places preferring
`resolved_symlink_dynamic_library` over `dynamic_library`.