Make BuildDriverKey non-CPU heavy.
Previously, a heavy-lifting part of conflict checking was done in the skyframe pool (collection of the action lookup values), it made sense to limit the concurrency by running it in the CPU-heavy pool. Now that we perform the entire conflict checking in an extra-Skyframe location, BuildDriverKey should no longer be CPU-heavy.

PiperOrigin-RevId: 561649161
Change-Id: Ic76661389d856ef9c492bd7b921649fa4455b583